+++ src/components/select/index.ts
@@ -6,3 +6,7 @@
  * actual or intended publication of such source code.
  */
 export {Select, Props as SelectProps} from './select';
+export {OptionLabel, Props as OptionLabelProps} from './option-label';
+export {SelectEmitter, Props as EmitterProps} from './select-emitter';
+export {Option, Props as OptionProps} from './option';
+export {useSelectContext} from './select-context';