+++ src/utils/use-constol-invalidity.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2019 Genestack Limited
+ * Copyright (c) 2011-2020 Genestack Limited
  * All Rights Reserved
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
  * The copyright notice above does not evidence any
@@ -8,10 +8,7 @@
 import * as React from 'react';
 
 /** Subscribe to native input invalidity */
-export function useInputInvalidity(
-    inputRef: React.RefObject<HTMLInputElement>,
-    invalid: boolean | undefined
-) {
+export function useInputInvalidity(inputRef: React.RefObject<HTMLInputElement>, invalid?: boolean) {
     // Input is valid by default
     const [invalidState, setInvalidState] = React.useState(invalid || false);