+++ changelog.md
@@ -2,8 +2,16 @@
 
 ## [Unreleased]
 
+## [10.0.0] - 2020-09-16
+
 ### Changed
 
+-   Changed [Select](src/components/select/select.tsx) API
+-   Add [Select](src/components/select/select.tsx) support in native and menu mode
+-   Add [Option](src/components/select/option.tsx)
+-   Add [OptionLabel](src/components/select/option-label.tsx)
+-   Add [SelectEmitter](src/components/select/select-emitter.tsx)
+-   Add [ArrowBottomThinIcon](src/icons/arrow-bottom-thin-icon.tsx)
 -   Changed a default font-family to `IBM Plex Sans`
 -   Changed `Typography` public properties:
     -   `quiet` to `intent = 'quiet'`
@@ -47,7 +55,7 @@
 
 ### Changed
 
--   Chnaged `<Paper />` to consume dark context and change own appearance depending on its value
+-   Changed `<Paper />` to consume dark context and change own appearance depending on its value
 -   Updated scrollbar colors in `<ScrollView />`
 
 ## [8.4.1] - 2020-06-30